size

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int b;
 long int c;
 long long int d;
 float f;
-printf("size of a char: %u byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of an int: %u byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %u byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %u byte(s)\n", (unsigned long)sizeof(d));
-printf("Size of a float: %u byte(s)\n", (unsigned long)sizeof(f));
+printf("size of a char: %u byte(s)\n", sizeof(a));
+printf("Size of an int: %u byte(s)\n", sizeof(b));
+printf("Size of a long int: %u byte(s)\n", sizeof(c));
+printf("Size of a long long int: %u byte(s)\n", sizeof(d));
+printf("Size of a float: %u byte(s)\n", sizeof(f));
 return (0);
 }
